+++ darts/models/forecasting/nlinear.py
@@ -12,7 +12,7 @@ from darts.models.forecasting.pl_forecasting_module import (
     io_processor,
 )
 from darts.models.forecasting.torch_forecasting_model import MixedCovariatesTorchModel
-from darts.utils.data.utils import PLModuleInput, TorchTrainingSample
+from darts.utils.data.torch_datasets.utils import PLModuleInput, TorchTrainingSample
 
 
 class _NLinearModule(PLForecastingModule):