+++ darts/models/forecasting/pl_forecasting_module.py
@@ -17,7 +17,7 @@ from joblib import Parallel, delayed
 from darts.logging import get_logger, raise_if, raise_log
 from darts.models.components.layer_norm_variants import RINorm
 from darts.timeseries import TimeSeries
-from darts.utils.data.utils import (
+from darts.utils.data.torch_datasets.utils import (
     PLModuleInput,
     TorchBatch,
     TorchInferenceBatch,