+++ darts/models/forecasting/tcn_model.py
@@ -19,7 +19,7 @@ from darts.models.forecasting.pl_forecasting_module import (
 from darts.models.forecasting.torch_forecasting_model import PastCovariatesTorchModel
 from darts.timeseries import TimeSeries
 from darts.utils.data import ShiftedTorchTrainingDataset, TorchTrainingDataset
-from darts.utils.data.utils import PLModuleInput, TorchTrainingSample
+from darts.utils.data.torch_datasets.utils import PLModuleInput, TorchTrainingSample
 from darts.utils.torch import MonteCarloDropout
 
 logger = get_logger(__name__)