+++ darts/utils/data/utils.py
@@ -1,10 +1,8 @@
-from collections.abc import Sequence
 from enum import Enum
-from typing import Optional, Union
+from typing import Union
 
 import numpy as np
 import pandas as pd
-import torch
 
 from darts import TimeSeries
 from darts.logging import get_logger, raise_log
@@ -16,74 +14,6 @@ logger = get_logger(__name__)
 DIVISIBLE_FREQS = {"D", "h", "H", "T", "min", "s", "S", "L", "ms", "U", "us", "N", "ns"}
 # supported built-in sample weight generators for regression and torch models
 SUPPORTED_SAMPLE_WEIGHT = {"linear", "exponential"}
-
-
-# `TorchTrainingDataset` output
-# (past target, past cov, historic future cov, future cov, static cov, sample weight, future target)
-TorchTrainingDatasetOutput = tuple[
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    np.ndarray,
-]
-TorchTrainingBatch = tuple[
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    torch.Tensor,
-]
-
-# `TorchInferenceDataset` output
-# (past target, past cov, future past cov, historic future cov, future cov, static cov, target series, pred time)
-TorchInferenceDatasetOutput = tuple[
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    TimeSeries,
-    Union[pd.Timestamp, int],
-]
-TorchInferenceBatch = tuple[
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    Sequence[TimeSeries],
-    Union[Sequence[pd.Timestamp], Sequence[int]],
-]
-
-TorchBatch = tuple[
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-    Optional[torch.Tensor],
-]
-
-# training sample has no sample weight
-# (past target, past cov, historic future cov, future cov, static cov, future target)
-TorchTrainingSample = tuple[
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    np.ndarray,
-]
-
-# the module receives three tensors
-# (past features (past target + past cov + historic future cov), future cov, static cov)
-PLModuleInput = tuple[torch.Tensor, Optional[torch.Tensor], Optional[torch.Tensor]]
 
 
 class FeatureType(Enum):
@@ -98,7 +28,6 @@ class FeatureType(Enum):
 
 
 # for extracting feature index boundaries
-_SampleIndexType = dict[FeatureType, tuple[Optional[int], Optional[int]]]
 
 _SERIES_TYPES = [
     FeatureType.PAST_TARGET,